chore: revert unrelated changes

+++ src/categories/categories.controller.ts
@@ -6,45 +6,43 @@ import {
   Patch,
   Param,
   Delete,
-  // UseGuards,
-  // Req,
-  // ForbiddenException,
+  UseGuards,
+  Req,
+  ForbiddenException,
+  HttpCode
 } from '@nestjs/common';
 import { CategoriesService } from './categories.service';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
 import { Category } from './entities/category.entity';
-// import { Role } from 'src/users/enum';
-// import { JwtAccessGuard } from 'src/auth/guards/jwt-access.guard';
+import { Role } from 'src/users/enum';
+import { JwtAccessGuard } from 'src/auth/guards/jwt-access.guard';
+import { ApiTags } from '@nestjs/swagger';
 
-// type AuthRequest = Request & {
-//   user: {
-//     id: string;
-//     role: Role;
-//   };
-// };
+type AuthRequest = Request & {
+  user: {
+    id: string;
+    role: Role;
+  };
+};
 
+@ApiTags('categories')
 @Controller('categories')
 export class CategoriesController {
   constructor(private readonly categoriesService: CategoriesService) {}
 
-  // @Post()
-  // @UseGuards(JwtAccessGuard)
-  // create(
-  //   @Body() createCategoryDto: CreateCategoryDto,
-  //   @Req() req: AuthRequest,
-  // ): Promise<Category> {
-  //   const user = req.user;
-
-  //   if (user.role !== Role.Admin) {
-  //     throw new ForbiddenException('Недостаточно прав');
-  //   }
-
-  //   return this.categoriesService.create(createCategoryDto);
-  // }
-
   @Post()
-  create(@Body() createCategoryDto: CreateCategoryDto): Promise<Category> {
+  @UseGuards(JwtAccessGuard)
+  create(
+    @Body() createCategoryDto: CreateCategoryDto,
+    @Req() req: AuthRequest,
+  ): Promise<Category> {
+    const user = req.user;
+
+    if (user.role !== Role.Admin) {
+      throw new ForbiddenException('Недостаточно прав');
+    }
+
     return this.categoriesService.create(createCategoryDto);
   }
 
@@ -59,15 +57,33 @@ export class CategoriesController {
   }
 
   @Patch(':id')
+  @UseGuards(JwtAccessGuard)
   update(
     @Param('id') id: string,
     @Body() updateCategoryDto: UpdateCategoryDto,
-  ) {
+    @Req() req: AuthRequest,
+  ): Promise<Category> {
+    const user = req.user;
+
+    if (user.role !== Role.Admin) {
+      throw new ForbiddenException('Недостаточно прав');
+    }
+
     return this.categoriesService.update(id, updateCategoryDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  @UseGuards(JwtAccessGuard)
+  @HttpCode(204)
+  async remove(
+    @Param('id') id: string,
+    @Req() req: AuthRequest,
+  ): Promise<void> {
+    const user = req.user;
+
+    if (user.role !== Role.Admin) {
+      throw new ForbiddenException('Недостаточно прав');
+    }
     return this.categoriesService.remove(id);
   }
 }
